requirement.txt
pygame==2.6.1
opencv-python==4.9.0.80
mediapipe==0.10.11
python-chess==1.10.0
numpy==1.26.4